pyproject.toml
[project]
name = "sucks"
dynamic = ["version"]
dependencies = [
    "PyYAML",
    "pydantic",
    "podman"
]


[project.scripts]
sucks="sucks.main:main"

[tool.setuptools.dynamic]
version = {attr = "sucks.__version__"}
readme = {file=["README.md"]}